Delete accounts app, move all logics to 'catalog'
