--- accounts/admin.py
@@ -1,8 +1,0 @@
-from django.contrib import admin
-
-from accounts import models
-
-
-@admin.register(models.User)
-class CustomUserAdmin(admin.ModelAdmin):
-    pass
