--- accounts/forms.py
@@ -1,28 +1,0 @@
-from django import forms
-from django.contrib.auth import get_user_model
-from django.contrib.auth.forms import UserCreationForm
-
-
-class LoginForm(forms.Form):
-    username = forms.CharField(
-        max_length=255,
-        widget=forms.TextInput(
-            attrs={
-                "placeholder": "Enter a valid username"
-            }
-        )
-    )
-    password = forms.CharField(
-        max_length=255,
-        widget=forms.PasswordInput(
-            attrs={
-                "placeholder": "Enter password"
-            }
-        )
-    )
-
-
-class RegistrationForm(UserCreationForm):
-    class Meta:
-        model = get_user_model()
-        fields = ("username", "first_name", "last_name", "email", "city")
