--- accounts/urls.py
@@ -1,12 +1,0 @@
-from django.urls import path
-
-from accounts import views
-
-
-app_name = "accounts"
-
-urlpatterns = [
-    path("login/", views.login_view, name="login"),
-    path("logout/", views.logout_view, name="logout"),
-    path("register/", views.UserCreateView.as_view(), name="register"),
-]
