+++ bonbon_shop/urls.py
@@ -23,7 +23,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path("accounts/", include("accounts.urls", namespace="accounts")),
     path("", include("catalog.urls", namespace="catalog")),
     path("__debug__/", include("debug_toolbar.urls")),
 ]
