+++ catalog/admin.py
@@ -1,6 +1,12 @@
 from django.contrib import admin
+from django.contrib.auth import get_user_model
 
 from catalog.models import Category, Product, Order, OrderItem
+
+
+@admin.register(get_user_model())
+class CustomUserAdmin(admin.ModelAdmin):
+    pass
 
 
 @admin.register(Category)
