+++ templates/accounts/login.html
@@ -9,14 +9,14 @@
     <h1>Login</h1>
     <div class="card">
       <div class="card-body">
-        <form action="" method="post">
+        <form action="" method="post" novalidate>
           {% csrf_token %}
           {{ form|crispy }}
           <div class="d-flex align-items-center justify-content-between">
             <input class="btn btn-glass" type="submit" value="Login">
             <div>
             <p class="d-inline">Don't have an account?</p>
-            <a href="{% url 'accounts:register' %}">Register</a>
+            <a href="{% url 'catalog:register' %}">Register</a>
             </div>
           </div>
         </form>
