+++ templates/accounts/register.html
@@ -15,7 +15,7 @@
             <input class="btn btn-glass" type="submit" value="Register">
             <div>
             <p class="d-inline">Already have an account?</p>
-            <a href="{% url 'accounts:login' %}">Login</a>
+            <a href="{% url 'catalog:login' %}">Login</a>
             </div>
           </div>
         </form>
